+++ src/Resources/Resources/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Added support for Azure resources deployment in Bicep language
 * Added support for -QueryString parameter in Test-Az*Deployments cmdlets
 * Fixed issue with dynamic parameters when New-Az*Deployments is used with -QueryString